JDBC - Clean up teardown methods, move mysql-specific stored proc test into MySQLDriverTest
