+++ src/test/java/ortus/boxlang/runtime/jdbc/drivers/MySQLDriverTest.java
@@ -11,6 +11,7 @@ import java.sql.SQLException;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIf;
@@ -25,6 +26,7 @@ import ortus.boxlang.runtime.types.IStruct;
 import ortus.boxlang.runtime.types.Query;
 import ortus.boxlang.runtime.types.Struct;
 import ortus.boxlang.runtime.types.exceptions.DatabaseException;
+import tools.JDBCTestUtils;
 
 @EnabledIf( "tools.JDBCTestUtils#hasMySQLModule" )
 public class MySQLDriverTest extends AbstractDriverTest {
@@ -52,43 +54,29 @@ public class MySQLDriverTest extends AbstractDriverTest {
 		MySQLDriverTest.createGeneratedKeyTable( mysqlDatasource, setUpContext );
 
 		// @TODO: Move mysql-specific StoredProcTest here and uncomment this setup.
-		// setupStoredProcTests();
+		// setupStoredProcTests( setUpContext );
 	}
 
 	@AfterAll
 	public static void teardown() throws SQLException {
-		AbstractDriverTest.teardownTestDatasource( new ScriptingRequestBoxContext( instance.getRuntimeContext() ), mysqlDatasource );
+		IBoxContext tearDownContext = new ScriptingRequestBoxContext( instance.getRuntimeContext() );
+		JDBCTestUtils.dropTestTable( mysqlDatasource, tearDownContext, "company", true );
+		AbstractDriverTest.teardownTestDatasource( tearDownContext, mysqlDatasource );
 	}
 
 	/***
 	 * Set up stored procedure and supporting table for testing. See StoredProcTest.
 	 */
-	public static void setupStoredProcTests( BoxRuntime instance, IBoxContext setUpContext ) {
-		mysqlDatasource.execute(
-		    """
-		    CREATE DEFINER=`root`@`%` PROCEDURE `sp_multi_result_set` (IN `companyName` VARCHAR(255))   BEGIN
-		    	SELECT *
-		    	FROM company
-		    	WHERE name <> companyName
-		    	order by name asc;
-
-		    	SELECT *
-		    	FROM company
-		    	WHERE name <> companyName
-		    	order by name desc;
-		    END$$
-		    """,
-		    setUpContext
-		);
+	public static void setupStoredProcTests( IBoxContext context ) {
 		mysqlDatasource.execute(
 		    """
 		    CREATE TABLE `company` (
 		    `id` int NOT NULL,
 		    `name` text NOT NULL,
 		    `active` tinyint(1) NOT NULL DEFAULT '1'
-		    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+		    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 		    	""",
-		    setUpContext
+		    context
 		);
 		mysqlDatasource.execute(
 		    """
@@ -98,8 +86,29 @@ public class MySQLDriverTest extends AbstractDriverTest {
 		    (3, 'Sony', 1),
 		    (4, 'Microsoft', 1);
 		    	""",
-		    setUpContext
+		    context
 		);
+		// formatter:off
+		mysqlDatasource.execute(
+		    """
+		    DELIMITER //
+		    CREATE PROCEDURE sp_multi_result_set(IN companyName VARCHAR(255))
+		    BEGIN
+		        SELECT *
+		        FROM company
+		        WHERE name <> companyName
+		        order by name asc;
+
+		        SELECT *
+		        FROM company
+		        WHERE name <> companyName
+		        order by name desc;
+		    END //
+		    DELIMITER ;
+		    """,
+		    context
+		);
+		// @formatter:on
 	}
 
 	/**
@@ -277,6 +286,23 @@ public class MySQLDriverTest extends AbstractDriverTest {
 		assertThat( t.getMessage() ).contains( "Boom!" );
 		assertThat( t.getCause() ).isNotNull();
 		assertThat( t.getCause().getMessage() ).contains( "Boom!" );
+	}
+
+	@Disabled( "Stored procedure initialization needs work; is failing with a syntax error on the first line." )
+	@Test
+	public void testMultiResultSets() {
+		getInstance().executeSource(
+		    String.format(
+		        """
+		            storedproc dataSource = "%s" procedure = "sp_multi_result_set" {
+		            	procparam sqltype="varchar" value="SEGA";
+		            	procresult name="names_asc" resultSet=1 maxRows=1;
+		            	procresult name="names_desc" resultSet=2 maxRows=2;
+		            };
+		        """, getDatasourceName() ),
+		    getContext(), BoxSourceType.BOXTEMPLATE );
+		assertThat( getVariables().get( Key.of( "names_asc" ) ) ).isInstanceOf( Query.class );
+		assertThat( getVariables().get( Key.of( "names_desc" ) ) ).isInstanceOf( Query.class );
 	}
 
 }
